supabase/functions/resume-analyzer/requirements.txt
supabase-functions
fastapi
pymupdf
python-docx
spacy
rake-nltk
openai
nltk
python-multipart
